small fixes

+++ result_saver.cpp
@@ -21,7 +21,7 @@ void StreamSaver::update(ResultHolder calc_result) {
 }
 
 void PercentLogger::update(ResultHolder calc_result) {
-    int percent = 100 * calc_result->task_num / tasks_size_;
+    int percent = int(100 * calc_result->task_num / tasks_size_);
     if (percent > cur_percent_) {
         cur_percent_ = percent;
         out_ << cur_percent_ << '%' << endl;
